Added history logic
